+++ package.json
@@ -25,6 +25,7 @@
     "lightweight-charts": "^4.1.2",
     "material-react-table": "^2.6.1",
     "react": "^18.2.0",
+    "react-chartjs-2": "^5.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.21.3",
     "react-select": "^5.8.0"
