+++ src/components/Table/index.ts
@@ -1,0 +1,2 @@
+export { default as AdvanceTable } from "./AdvanceTable";
+export { default as SimpleTable } from "./SimpleTable";
