+++ src/components/index.ts
@@ -1,5 +1,5 @@
 export { default as About } from "./About";
-export { default as AdvanceTable } from "./AdvanceTable";
+export { SimpleTable, AdvanceTable } from "./Table";
 export { default as Breadcrumb } from "./Breadcrumb";
 export { default as Chart } from "./Chart";
 export { default as Contact } from "./Contact";
@@ -7,8 +7,10 @@ export { default as Footer } from "./Footer";
 export { default as ForgotPassword } from "./ForgotPassword";
 export { default as Login } from "./Login";
 export { default as Navbar } from "./Navbar";
+export { default as NotFound } from "./NotFound";
 export { default as PrivacyPolicy } from "./PrivacyPolicy";
 export { default as Services } from "./Services";
 export { default as ShareCalculator } from "./ShareCalculator";
 export { default as Signup } from "./Signup";
-export { default as SimpleTable } from "./SimpleTable";
+export { default as SimpleChart } from "./SimpleChart";
+export { default as StockDetail } from "./StockDetail";
